requirements.txt
mysql-connector-python==9.3.0
paho-mqtt==2.1.0
influxdb-client==1.41.0
